Rename generator modules
