+++ flatdata-py/generator/tests/generators/test_go_generator.py
@@ -3,14 +3,14 @@
  See the LICENSE file in the root of this project for license details.
 '''
 
-from generator.generators.GoGenerator import GoGenerator
+from generator.generators.go import GoGenerator
 
 from .assertions import *
 
 
 def test_imports_and_constants_generation():
     generate_and_assert_in("""namespace xyz{
-        
+
 }
     """, GoGenerator, """/////////////////////////////////////////////////////////////////////////
 //    ATTENTION!
@@ -115,7 +115,7 @@ func (v *AArchive) GetSizeInBytes() int {
 	}""",
         """// Add resources to archive
     v.R0Instance = &AR0Instance {
-        descriptor: r0MemoryDescriptor, 
+        descriptor: r0MemoryDescriptor,
         IsOptional: false,
         IsOpen: r0IsOpen,
     }
@@ -162,7 +162,7 @@ func (v *AVectorResourceVector) GetSize() int {
 }
 
 func (v *AVectorResourceVector) GetSlice(start, end, step int) []*T {
-	var result []*T	
+	var result []*T
 	for start <= end {
 		result = append(result, &T{
 			descriptor: v.descriptor,
@@ -180,7 +180,7 @@ func (v *AVectorResourceVector) Close() {
 
 func (v *AVectorResourceVector) GetSizeInBytes() int {
     return v.descriptor.Len()
-}""" 
+}"""
     ]
 
     generate_and_assert_in("""
@@ -233,7 +233,7 @@ func (v *AMultivectorResourceVector) GetSize() int {
 }
 
 func (v *AMultivectorResourceVector) GetSlice(start, end, step int) []*IndexType33 {
-	var result []*IndexType33	
+	var result []*IndexType33
 	for start <= end {
 		result = append(result, &IndexType33{
 			descriptor: v.descriptor,
@@ -276,7 +276,7 @@ func (v *AMultivectorResourceMultivector) GetSizeInBytes() int {
 func (v *AMultivectorResourceMultivector) getBucketOffset(i int) int {
 	if i == v.index.GetSize() {
 		return v.descriptor.Len() - int(flatdataPaddingSizeInBytes)
-	} 
+	}
 	return int(v.index.Get(i).GetValue()) + int(flatdataOffsetSizeInBytes)
 }
 
@@ -293,7 +293,7 @@ func (v *AMultivectorResourceMultivector) Get(i int) []interface{} {
 			//TODO: How to process case, then type of element is not found?
 			log.Println("Can't get type of element")
 		}
-		
+
 		switch element := abstractElement.(type) {
 		case *T:
 			element.position = offset
@@ -308,7 +308,7 @@ func (v *AMultivectorResourceMultivector) Get(i int) []interface{} {
 			log.Println("Can't cast element. Type is unknown...")
 		}
 	}
-	
+
 	return result
 }"""
     ]
@@ -420,7 +420,7 @@ func (v *ARawDataResourceRawData) GetSizeInBytes() int {
 		}
 	}""",
         """v.RawDataResourceRawData = &ARawDataResourceRawData{
-        descriptor: rawDataResourceMemoryDescriptor, 
+        descriptor: rawDataResourceMemoryDescriptor,
         IsOptional: true,
         IsOpen: rawDataResourceIsOpen,
     }"""
